feat(34683536): dynamic generation metadata for invitePage
